Add `with-magit-boost-buffer` macro
This commit introduces a new macro, `with-magit-boost-buffer`, to
simplify and standardize access to existing `magit-boost` buffers.

Previously, several functions contained repetitive `when-let` and
`with-current-buffer` constructs to locate and operate within a specific
Magit Boost buffer. This pattern led to some code duplication and
reduced readability.

The new `with-magit-boost-buffer` macro encapsulates this common logic.
It finds an existing boost buffer for a given directory and connection
type, and executes the provided body within that buffer if found.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>